+++ apps/api/package.json
@@ -15,11 +15,9 @@
     "@oyster/core": "workspace:*",
     "@oyster/types": "workspace:*",
     "@oyster/utils": "workspace:*",
-    "@sentry/node": "^7.43.0",
-    "@sentry/tracing": "^7.43.0",
+    "@sentry/bun": "^9.34.0",
     "cors": "^2.8.5",
     "dedent": "^0.7.0",
-    "dotenv": "^16.3.1",
     "express": "^4.18.3",
     "helmet": "^6.0.1",
     "puppeteer-core": "^23.7.0",
