--- apps/api/src/routers/health.router.ts
@@ -1,7 +1,0 @@
-import express from 'express';
-
-export const healthRouter = express.Router();
-
-healthRouter.get('/health', async (_, res) => {
-  return res.send({});
-});
