--- apps/api/src/shared/types.ts
@@ -1,3 +1,0 @@
-import type { Request } from 'express';
-
-export type RawBodyRequest = Request & { rawBody?: Buffer };
